refactor: AdminPage UI 분리
